refactor: revised logic for restoring saved data
revised logic for restoring store plant root weight data on tab back navigation

BREAKING CHANGE:
