+++ app/src/main/java/com/akilimo/rya/views/fragments/ui/TriangleFragment.kt
@@ -30,6 +30,9 @@ class TriangleFragment : BasePlantTriangle() {
     private var _binding: FragmentTriangleBinding? = null
     private var ctx: Context? = null
 
+
+    private val inputLayouts: MutableList<TextInputLayout> = arrayListOf()
+
     private val binding get() = _binding!!
 
     companion object {
@@ -68,13 +71,16 @@ class TriangleFragment : BasePlantTriangle() {
         super.onViewCreated(view, savedInstanceState)
         val lyt = binding.lytTextField
         lyt.removeAllViews() //clear all components
+        inputLayouts.clear()
         for (i in 0 until triangleCount) {
             val textInputLayout = addTextInputLayout(i, requireView().context)
             lyt.addView(textInputLayout)
             inputLayouts.add(textInputLayout)
         }
+    }
 
-        //get saved values and add them to text fields
+
+    override fun loadTriangleData() {
         var plantNumber = 1
         for (inputLayout in inputLayouts) {
             val plantTriangle = database?.plantTriangleDao()
@@ -86,11 +92,46 @@ class TriangleFragment : BasePlantTriangle() {
         }
     }
 
-
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
     }
+
+
+    override fun verifyStep(): VerificationError? {
+        var inputValid = false
+        val plantTrianglesMeasurement: MutableList<PlantTriangleEntity> = arrayListOf()
+        var plantNumber = 1
+        for (inputLayout in inputLayouts) {
+            val rootWeightString = inputLayout.editText?.editableText.toString()
+            val rootWeight = StringToNumberFactory.stringToDouble(rootWeightString)
+            inputValid = rootWeight > 0
+            if (inputValid) {
+                //save this value to the database
+                inputLayout.error = null
+                plantTrianglesMeasurement.add(
+                    PlantTriangleEntity(
+                        triangleName = triangleName!!,
+                        plantName = "plant$plantNumber",
+                        rootWeight = rootWeight
+                    )
+                )
+                plantNumber++
+            } else {
+                inputLayout.error = "Provide correct plant root weight"
+                inputLayout.requestFocus()
+                break //no need to loop all through
+            }
+        }
+
+        if (!inputValid) {
+            return VerificationError("Provide correct plant root weight for all inputs")
+        }
+
+        database?.plantTriangleDao()?.insertAll(plantTrianglesMeasurement)
+        return verificationError
+    }
+
 
     override fun onError(error: VerificationError) {
         val snackBar = Snackbar.make(
