+++ app/src/main/java/com/akilimo/rya/views/fragments/ui/TriangleTwoFragment.kt
@@ -14,6 +14,7 @@ import com.akilimo.rya.databinding.FragmentTriangleBinding
 import com.akilimo.rya.entities.PlantTriangleEntity
 import com.akilimo.rya.utils.StringToNumberFactory
 import com.google.android.material.snackbar.Snackbar
+import com.google.android.material.textfield.TextInputLayout
 import com.stepstone.stepper.VerificationError
 
 
@@ -30,6 +31,9 @@ class TriangleTwoFragment : BasePlantTriangle() {
     private var _binding: FragmentTriangleBinding? = null
     private var ctx: Context? = null
 
+
+    private val inputLayouts: MutableList<TextInputLayout> = arrayListOf()
+
     private val binding get() = _binding!!
 
     companion object {
@@ -38,11 +42,10 @@ class TriangleTwoFragment : BasePlantTriangle() {
          * @return A new instance of fragment TriangleFragment.
          */
         @JvmStatic
-        fun newInstance(triangleCount: Int, triangleName: String) =
-            TriangleTwoFragment().apply {
-                this.triangleName = triangleName
-                this.triangleCount = triangleCount
-            }
+        fun newInstance(triangleCount: Int, triangleName: String) = TriangleTwoFragment().apply {
+            this.triangleName = triangleName
+            this.triangleCount = triangleCount
+        }
 
         @JvmStatic
         fun newInstance() = TriangleTwoFragment().apply { }
@@ -55,38 +58,30 @@ class TriangleTwoFragment : BasePlantTriangle() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {
-            triangleCount = it.getInt(PLANT_COUNT)
-            triangleName = it.getString(TRIANGLE_NAME)
-        }
         database = AppDatabase.getDatabase(ctx!!)
     }
 
     override fun loadFragmentLayout(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ): View {
         _binding = FragmentTriangleBinding.inflate(inflater, container, false)
         return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         val lyt = binding.lytTextField
         lyt.removeAllViews() //clear all components
+        inputLayouts.clear()
         for (i in 0 until triangleCount) {
             val textInputLayout = addTextInputLayout(i, requireView().context)
             lyt.addView(textInputLayout)
             inputLayouts.add(textInputLayout)
         }
 
-        //get saved values and add them to text fields
+    }
+
+    override fun loadTriangleData() {
         var plantNumber = 1
         for (inputLayout in inputLayouts) {
             val plantTriangle = database?.plantTriangleDao()
@@ -98,10 +93,49 @@ class TriangleTwoFragment : BasePlantTriangle() {
         }
     }
 
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    override fun verifyStep(): VerificationError? {
+        var inputValid = false
+        val plantTrianglesMeasurement: MutableList<PlantTriangleEntity> = arrayListOf()
+        var plantNumber = 1
+        for (inputLayout in inputLayouts) {
+            val rootWeightString = inputLayout.editText?.editableText.toString()
+            val rootWeight = StringToNumberFactory.stringToDouble(rootWeightString)
+            inputValid = rootWeight > 0
+            if (inputValid) {
+                //save this value to the database
+                inputLayout.error = null
+                plantTrianglesMeasurement.add(
+                    PlantTriangleEntity(
+                        triangleName = triangleName!!,
+                        plantName = "plant$plantNumber",
+                        rootWeight = rootWeight
+                    )
+                )
+                plantNumber++
+            } else {
+                inputLayout.error = "Provide correct plant root weight"
+                inputLayout.requestFocus()
+                break //no need to loop all through
+            }
+        }
+
+        if (!inputValid) {
+            return VerificationError("Provide correct plant root weight for all inputs")
+        }
+
+        database?.plantTriangleDao()?.insertAll(plantTrianglesMeasurement)
+        return verificationError
+    }
+
+
     override fun onError(error: VerificationError) {
         val snackBar = Snackbar.make(
-            binding.constraintLayout, error.errorMessage,
-            Snackbar.LENGTH_SHORT
+            binding.constraintLayout, error.errorMessage, Snackbar.LENGTH_SHORT
         )
 
         snackBar.setAction("RETRY") {
